requirements.txt
djangorestframework==3.12.4
psycopg2==2.9.1
